Refractored the maze so it stores cells
Refractored the maze so it stores Cell class objects and inside them objects of different types of Object classes. Made it so player can walk around and win the game if he walks out of bounds of 2d array. Also implemented so if player walks on a monster the monster dissapears (fighting mechanics will be implemented soon)